Two Things
1. Added KODS Neutron

Neutron Parts are very WIP, performance is being tweaked, but this is enough to start, comes with a WIP Archimedes Engine Config.

2. Removed BoringCrewServices

BCS is now officially included in RealismOverhaul, there is no need for this repo to have the configs now, please download those configs with the official RO release.